💥 core,zlink: chain::ReplyStream's items now owned
This was a major fuckup from my side that I thought that the Stream impl
of chain::ReplayStream can return borrowed items. This means that
replies can be over-written while still borrowed by the user.

Unfortunately, the Stream API doesn't currently allow us to borrow from
self, so we have no choice but to switch to require DeserializeOwned for
the item type.

In the future, this may be solvable when Stream allows borrowing from
self.

Fixes #185.